requirements.txt
fastapi
pydantic
starlette
typing-extensions
# uvicorn
uvicorn[standard]
gunicorn
requests
python-dotenv
pytz
python-multipart
SQLAlchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
# GeoAlchemy2
logzio-python-handler
aiofiles
jinja2
fastapi-restful[tasks]
numpy
pandas
google
protobuf==3.19.*
GeoAlchemy2
shapely
websockets
fastapi-socketio
redis_om
fastapi-redis-cache
fastapi-cache2[redis]
aioredis
versiontag
gitpython
crython
pickle